Added a simple GUI menu using the final FOSS version of PySimpleGUI
Fix local CSV sync to use quick check

The "Sync from Local CSV" option was incorrectly performing a full,
thorough check by default. This was unintuitive, as the expectation
is for it to quickly add missing workouts from the backup file.

This change updates the call to sync_workouts to use `full_check=False`,
aligning its behavior with the "Quick Sync" philosophy.